commit 17abec598417b0120193b3ad4e5390d5d8682014
    12385 tst.subr.d needs fixing for NULL as a pointer
commit 584b574a3b16c6772c8204ec1d1c957c56f22a87
    12174 i86pc: variable may be used uninitialized
commit 34a4e6b53d2e6f2605fd77cda5b161201d7d0f20
    12377 aggr: may be used uninitialized
commit c9ffe217655ce62448bdb45d6f113f43b4dfcc54
    12367 aac: variable may be used uninitialized
commit a25e615d76804404e5fc63897a9196d4f92c3f5e
    12371 dis x86 EVEX prefix mishandled
    12372 dis EVEX encoding SIB mishandled
    12373 dis support for EVEX vaes instructions
    12374 dis support for EVEX vpclmulqdq instructions
    12375 dis support for gfni instructions
commit c1e9bf00765d7ac9cf1986575e4489dd8710d9b1
    12369 dis WBNOINVD support

	Conflicts:
	usr/src/common/dis/i386/dis_tables.c
	exception_lists/wscheck

+++ usr/src/cmd/dtrace/test/tst/common/aggs/tst.subr.d
@@ -22,7 +22,7 @@
 /*
  * Copyright 2007 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
- * Copyright (c) 2012, Joyent, Inc.  All rights reserved.
+ * Copyright 2020 Joyent, Inc.
  */
 
 #include <sys/dtrace.h>
@@ -61,8 +61,8 @@ INTFUNC(mutex_type_spin(&`cpu_lock))
 INTFUNC(rw_read_held(&`vfssw_lock))
 INTFUNC(rw_write_held(&`vfssw_lock))
 INTFUNC(rw_iswriter(&`vfssw_lock))
-INTFUNC(copyin(NULL, 1))
-STRFUNC(copyinstr(NULL, 1))
+INTFUNC(copyin((uintptr_t)NULL, 1))
+STRFUNC(copyinstr((uintptr_t)NULL, 1))
 INTFUNC(speculation())
 INTFUNC(progenyof($pid))
 INTFUNC(strlen("fooey"))
